+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.11.0 (2025-09-08)
 - 🧑‍💻 Improve helping messages in settings
+- 🐛 Fixed type in "Erase all groups" message
 
 ## v1.10.0 (2025-09-05)
 - ✨ Added copy and cut tabs from a group into a different group [#39](https://github.com/tab-guardian/tab-guardian/issues/39)